Move ReplAutocompletion to autocompletion package (#306)
